+++ src/main/java/com/yoda/item/service/ItemService.java
@@ -5,21 +5,21 @@ import java.util.List;
 import com.yoda.item.model.Item;
 
 public interface ItemService {
-	void save(Long siteId, Item item);
+	void save(int siteId, Item item);
 
 	Item getItem(int itemId);
 
-	Item getItem(long siteId, int itemId);
+	Item getItem(int siteId, int itemId);
 
-	Item getItem(long siteId, String itemNaturalKey);
+	Item getItem(int siteId, String itemNaturalKey);
 
 	Item update(Item item);
 
 	Item update(int id, String brand, String description, String level, String name, int price);
 
-	List<Item> getItems(long siteId);
+	List<Item> getItems(int siteId);
 
-	List<Item> search(long siteId, String itemNum, String itemUpcCd, String itemShortDesc);
+	List<Item> search(int siteId, String itemNum, String itemUpcCd, String itemShortDesc);
 
 	void remove(int itemId);
